Refactored AbstractGoogleAuthActivity and MainActivity to provide loginViewModel
